more

+++ mage_ai/data_preparation/variable_manager.py
@@ -4,9 +4,11 @@ from typing import Any, Dict, List, Optional, Union
 
 from mage_ai.data.constants import InputDataType
 from mage_ai.data.tabular.models import BatchSettings
+from mage_ai.data.variables.constants import GLOBAL_DIRECTORY_NAME
 from mage_ai.data_preparation.models.block.settings.variables.models import (
     ChunkKeyTypeUnion,
 )
+from mage_ai.data_preparation.models.constants import PIPELINES_FOLDER
 from mage_ai.data_preparation.models.utils import (
     infer_variable_type,
     warn_for_repo_path,
@@ -25,7 +27,12 @@ from mage_ai.system.models import ResourceUsage
 
 
 class VariableManager:
-    def __init__(self, repo_path=None, variables_dir=None):
+    def __init__(
+        self,
+        repo_path: Optional[str] = None,
+        variables_dir: Optional[str] = None,
+        pipeline_uuid: Optional[str] = None,
+    ):
         warn_for_repo_path(repo_path)
         self.repo_path = repo_path or get_repo_path()
         if variables_dir is None:
@@ -33,6 +40,7 @@ class VariableManager:
         else:
             self.variables_dir = variables_dir
         self.storage = LocalStorage()
+        self.pipeline_uuid = pipeline_uuid
         # TODO: implement caching logic
 
     @classmethod
@@ -40,10 +48,12 @@ class VariableManager:
         cls,
         repo_path: Optional[str] = None,
         variables_dir: Optional[str] = None,
+        pipeline_uuid: Optional[str] = None,
     ) -> 'VariableManager':
         manager_args = dict(
             repo_path=repo_path,
             variables_dir=variables_dir,
+            pipeline_uuid=pipeline_uuid,
         )
         if variables_dir is not None and variables_dir.startswith(S3_PREFIX):
             return S3VariableManager(**manager_args)
@@ -54,21 +64,24 @@ class VariableManager:
 
     def add_variable(
         self,
-        pipeline_uuid: str,
         block_uuid: str,
         variable_uuid: str,
         data: Any,
-        partition: Optional[str] = None,
-        variable_type: Optional[VariableType] = None,
         clean_block_uuid: bool = True,
+        clean_variable_uuid: bool = True,
         disable_variable_type_inference: bool = False,
+        partition: Optional[str] = None,
+        persist: bool = True,
+        pipeline_uuid: Optional[str] = None,
+        variable_type: Optional[VariableType] = None,
+        # Data reader and writer settings
         input_data_types: Optional[List[InputDataType]] = None,
         resource_usage: Optional[ResourceUsage] = None,
         read_batch_settings: Optional[BatchSettings] = None,
         read_chunks: Optional[List[ChunkKeyTypeUnion]] = None,
         write_batch_settings: Optional[BatchSettings] = None,
         write_chunks: Optional[List[ChunkKeyTypeUnion]] = None,
-    ) -> None:
+    ) -> Variable:
         """
         Used by:
             block
@@ -82,13 +95,14 @@ class VariableManager:
             )
 
         variable = Variable(
-            clean_name(variable_uuid),
+            clean_name(variable_uuid) if clean_variable_uuid else variable_uuid,
             self.pipeline_path(pipeline_uuid),
             block_uuid,
+            clean_block_uuid=clean_block_uuid,
             partition=partition,
             storage=self.storage,
             variable_type=variable_type,
-            clean_block_uuid=clean_block_uuid,
+            # Data reader and writer settings
             input_data_types=input_data_types,
             resource_usage=resource_usage,
             read_batch_settings=read_batch_settings,
@@ -97,51 +111,31 @@ class VariableManager:
             write_chunks=write_chunks,
         )
 
-        # Delete data if it exists
-        variable.delete()
-        variable.variable_type = variable_type
-        variable.write_data(data)
+        if persist:
+            # Delete data if it exists
+            variable.delete()
+            variable.variable_type = variable_type
+            variable.write_data(data)
 
-        if is_debug():
-            print(
-                f'Variable {variable_uuid} ({variable_type or "no type"}) for block {block_uuid} '
-                f"in pipeline {pipeline_uuid} "
-                f"stored in {variable.variable_path}"
-            )
+            if is_debug():
+                print(
+                    f'Variable {variable_uuid} ({variable_type or "no type"}) for block '
+                    f'{block_uuid} stored in {variable.variable_path}'
+                )
 
-    def build_variable(
-        self,
-        pipeline_uuid: str,
-        block_uuid: str,
-        variable_uuid: str,
-        partition: Optional[str] = None,
-        variable_type: Optional[VariableType] = None,
-        clean_variable_uuid: bool = True,
-        input_data_types: Optional[List[InputDataType]] = None,
-        read_batch_settings: Optional[BatchSettings] = None,
-        read_chunks: Optional[List[ChunkKeyTypeUnion]] = None,
-    ) -> Variable:
-        return Variable(
-            clean_name(variable_uuid) if clean_variable_uuid else variable_uuid,
-            self.pipeline_path(pipeline_uuid),
-            block_uuid,
-            partition=partition,
-            storage=self.storage,
-            variable_type=variable_type,
-            read_batch_settings=read_batch_settings,
-            read_chunks=read_chunks,
-            input_data_types=input_data_types,
-        )
+        return variable
 
     async def add_variable_async(
         self,
-        pipeline_uuid: str,
         block_uuid: str,
         variable_uuid: str,
         data: Any,
-        partition: Optional[str] = None,
-        variable_type: Optional[VariableType] = None,
         clean_block_uuid: bool = True,
+        clean_variable_uuid: bool = True,
+        partition: Optional[str] = None,
+        pipeline_uuid: Optional[str] = None,
+        variable_type: Optional[VariableType] = None,
+        # Data reader and writer settings
         input_data_types: Optional[List[InputDataType]] = None,
         resource_usage: Optional[ResourceUsage] = None,
         read_batch_settings: Optional[BatchSettings] = None,
@@ -160,13 +154,14 @@ class VariableManager:
         )
 
         variable = Variable(
-            clean_name(variable_uuid),
+            clean_name(variable_uuid) if clean_variable_uuid else variable_uuid,
             self.pipeline_path(pipeline_uuid),
             block_uuid,
+            clean_block_uuid=clean_block_uuid,
             partition=partition,
             storage=self.storage,
             variable_type=variable_type,
-            clean_block_uuid=clean_block_uuid,
+            # Data reader and writer settings
             input_data_types=input_data_types,
             resource_usage=resource_usage,
             read_batch_settings=read_batch_settings,
@@ -246,89 +241,67 @@ class VariableManager:
 
     def get_variable(
         self,
-        pipeline_uuid: str,
-        block_uuid: str,
         variable_uuid: str,
+        block_uuid: Optional[str] = None,
+        clean_block_uuid: bool = True,
         dataframe_analysis_keys: Optional[List[str]] = None,
         partition: Optional[str] = None,
-        variable_type: Optional[VariableType] = None,
-        raise_exception: bool = False,
+        pipeline_uuid: Optional[str] = None,
+        raise_exception: Optional[bool] = None,
+        read_data: bool = True,
         sample: bool = False,
         sample_count: Optional[int] = None,
         spark: Optional[Any] = None,
-        clean_block_uuid: bool = True,
-        input_data_types: Optional[List[InputDataType]] = None,
-        read_batch_settings: Optional[BatchSettings] = None,
-        read_chunks: Optional[List[ChunkKeyTypeUnion]] = None,
-        write_batch_settings: Optional[BatchSettings] = None,
-        write_chunks: Optional[List[ChunkKeyTypeUnion]] = None,
-    ) -> Any:
-        variable = self.get_variable_object(
-            pipeline_uuid,
-            block_uuid,
-            variable_uuid,
-            partition=partition,
-            variable_type=variable_type,
-            spark=spark,
-            clean_block_uuid=clean_block_uuid,
-            input_data_types=input_data_types,
-            read_batch_settings=read_batch_settings,
-            read_chunks=read_chunks,
-            write_batch_settings=write_batch_settings,
-            write_chunks=write_chunks,
-        )
-        return variable.read_data(
-            dataframe_analysis_keys=dataframe_analysis_keys,
-            raise_exception=raise_exception,
-            sample=sample,
-            sample_count=sample_count,
-            spark=spark,
-        )
-
-    def get_variable_object(
-        self,
-        pipeline_uuid: str,
-        block_uuid: str,
-        variable_uuid: str,
-        partition: Optional[str] = None,
         variable_type: Optional[VariableType] = None,
-        clean_block_uuid: bool = True,
-        spark=None,
+        # Data reader and writer settings
         input_data_types: Optional[List[InputDataType]] = None,
         read_batch_settings: Optional[BatchSettings] = None,
         read_chunks: Optional[List[ChunkKeyTypeUnion]] = None,
         write_batch_settings: Optional[BatchSettings] = None,
         write_chunks: Optional[List[ChunkKeyTypeUnion]] = None,
-    ) -> Variable:
-        if variable_type == VariableType.DATAFRAME and spark is not None:
+    ) -> Union[Any, Variable]:
+        if spark is not None and variable_type == VariableType.DATAFRAME:
             variable_type = VariableType.SPARK_DATAFRAME
-        return Variable(
+
+        variable = Variable(
             variable_uuid,
             self.pipeline_path(pipeline_uuid),
-            block_uuid,
+            block_uuid or GLOBAL_DIRECTORY_NAME,
+            clean_block_uuid=clean_block_uuid,
             partition=partition,
             spark=spark,
             storage=self.storage,
             variable_type=variable_type,
-            clean_block_uuid=clean_block_uuid,
+            # Data reader and writer settings
             input_data_types=input_data_types,
             read_batch_settings=read_batch_settings,
             read_chunks=read_chunks,
             write_batch_settings=write_batch_settings,
             write_chunks=write_chunks,
         )
+
+        if read_data:
+            return variable.read_data(
+                dataframe_analysis_keys=dataframe_analysis_keys,
+                raise_exception=False if raise_exception is None else raise_exception,
+                sample=sample,
+                sample_count=sample_count,
+                spark=spark,
+            )
+
+        return variable
 
     def get_variables_by_pipeline(self, pipeline_uuid: str) -> Dict[str, List[str]]:
         from mage_ai.data_preparation.models.pipeline import Pipeline
 
         pipeline = Pipeline.get(pipeline_uuid, repo_path=self.repo_path)
-        variable_dir_path = os.path.join(self.pipeline_path(pipeline_uuid), VARIABLE_DIR)
+        variable_dir_path = self.variable_dir_path(pipeline_uuid=pipeline_uuid)
         if not self.storage.path_exists(variable_dir_path):
             return dict()
         block_dirs = self.storage.listdir(variable_dir_path)
         variables_by_block = dict()
         for d in block_dirs:
-            if not pipeline.has_block(d) and d != 'global':
+            if not pipeline.has_block(d) and d != GLOBAL_DIRECTORY_NAME:
                 continue
             block_variables_path = os.path.join(variable_dir_path, d)
             if not self.storage.isdir(block_variables_path):
@@ -339,54 +312,97 @@ class VariableManager:
                 variables_by_block[d] = [v for v in variable_names if v != '']
         return variables_by_block
 
-    def get_variables_by_block(
+    def get_variable_uuids(
         self,
-        pipeline_uuid: str,
-        block_uuid: str,
-        partition: Optional[str] = None,
+        block_uuid: Optional[str] = None,
         clean_block_uuid: bool = True,
         max_results: Optional[int] = None,
-    ) -> List[str]:
-        variable_dir_path = os.path.join(
-            self.pipeline_path(pipeline_uuid),
-            VARIABLE_DIR,
-            partition or '',
-            clean_name(block_uuid) if clean_block_uuid else block_uuid,
+        partition: Optional[str] = None,
+        pipeline_uuid: Optional[str] = None,
+    ) -> List[Variable]:
+        variable_dir_path = self.variable_dir_path(
+            block_uuid=block_uuid or GLOBAL_DIRECTORY_NAME,
+            clean=clean_block_uuid,
+            partition=partition,
+            pipeline_uuid=pipeline_uuid,
         )
         if not self.storage.path_exists(variable_dir_path):
             return []
+
         opts = {}
         if max_results is not None:
             opts['max_results'] = max_results
-        variables = self.storage.listdir(variable_dir_path, **opts)
-        variables = [v for v in variables if v.split('.')[0]]
+
+        variable_uuids = self.storage.listdir(variable_dir_path, **opts)
+        variable_uuids = [v for v in variable_uuids if v.split('.')[0]]
 
         def __sort(variable_uuid: str) -> List[Union[int, str]]:
             key = variable_uuid.split('.')[0]
             parts = key.split('_')
             return [int(k) if k.isdigit() else k for k in parts]
 
-        return sorted(variables, key=__sort)
+        return [
+            Variable(
+                variable_uuid,
+                self.pipeline_path(pipeline_uuid),
+                block_uuid or GLOBAL_DIRECTORY_NAME,
+            )
+            for variable_uuid in sorted(variable_uuids, key=__sort)
+        ]
 
-    def pipeline_path(self, pipeline_uuid: str) -> str:
-        path = os.path.join(self.variables_dir, 'pipelines', pipeline_uuid)
+    def pipeline_path(self, pipeline_uuid: Optional[str] = None) -> str:
+        path = os.path.join(
+            self.variables_dir,
+            PIPELINES_FOLDER,
+            pipeline_uuid or self.pipeline_uuid or '',
+        )
         if type(self.storage) is LocalStorage:
             if not self.storage.path_exists(path):
                 self.storage.makedirs(path, exist_ok=True)
         return path
 
+    def variable_dir_path(
+        self,
+        block_uuid: Optional[str] = None,
+        clean: bool = True,
+        partition: Optional[str] = None,
+        pipeline_uuid: Optional[str] = None,
+        variable_uuid: Optional[str] = None,
+    ) -> str:
+        return os.path.join(
+            self.pipeline_path(pipeline_uuid),
+            VARIABLE_DIR,
+            partition or '',
+            (clean_name(block_uuid) if clean else block_uuid) if block_uuid else '',
+            variable_uuid or '',
+        )
+
 
 class S3VariableManager(VariableManager):
-    def __init__(self, repo_path=None, variables_dir=None):
-        super().__init__(repo_path=repo_path, variables_dir=variables_dir)
+    def __init__(
+        self,
+        repo_path: Optional[str] = None,
+        variables_dir: Optional[str] = None,
+        pipeline_uuid: Optional[str] = None,
+    ):
+        super().__init__(
+            repo_path=repo_path, variables_dir=variables_dir, pipeline_uuid=pipeline_uuid
+        )
         from mage_ai.data_preparation.storage.s3_storage import S3Storage
 
         self.storage = S3Storage(dirpath=variables_dir)
 
 
 class GCSVariableManager(VariableManager):
-    def __init__(self, repo_path=None, variables_dir=None):
-        super().__init__(repo_path=repo_path, variables_dir=variables_dir)
+    def __init__(
+        self,
+        repo_path: Optional[str] = None,
+        variables_dir: Optional[str] = None,
+        pipeline_uuid: Optional[str] = None,
+    ):
+        super().__init__(
+            repo_path=repo_path, variables_dir=variables_dir, pipeline_uuid=pipeline_uuid
+        )
         from mage_ai.data_preparation.storage.gcs_storage import GCSStorage
 
         self.storage = GCSStorage(dirpath=variables_dir)
@@ -420,11 +436,10 @@ def get_global_variables(
     else:
         variables_dir = variables_dir or get_variables_dir()
         variables = VariableManager(
-            repo_path=repo_path, variables_dir=variables_dir
-        ).get_variables_by_block(
-            pipeline_uuid,
-            'global',
-        )
+            repo_path=repo_path,
+            variables_dir=variables_dir,
+            pipeline_uuid=pipeline_uuid,
+        ).get_variable_uuids()
         global_variables = dict()
         for variable in variables:
             global_variables[variable] = get_global_variable(pipeline_uuid, variable)
@@ -448,11 +463,9 @@ def get_global_variable(
     if pipeline.variables is not None:
         return pipeline.variables.get(key)
     else:
-        return VariableManager(variables_dir=get_variables_dir()).get_variable(
-            pipeline_uuid,
-            'global',
-            key,
-        )
+        return VariableManager(
+            variables_dir=get_variables_dir(), pipeline_uuid=pipeline_uuid
+        ).get_variable(variable_uuid=key)
 
 
 def get_variable(pipeline_uuid: str, block_uuid: str, key: str, **kwargs) -> Any:
@@ -504,4 +517,4 @@ def delete_global_variable(
     else:
         VariableManager(
             variables_dir=get_variables_dir(),
-        ).delete_variable(pipeline_uuid, 'global', key)
+        ).delete_variable(pipeline_uuid, GLOBAL_DIRECTORY_NAME, key)
